fix: moved simple did key consistency tests to jvmTest folder
